Revert "Move mdx components to mdx-components.tsx"
This reverts commit 0d6d31d1f2cef2acde271f4defee88a671425a61.
