+++ packages/web/docs/mdx-components.js
@@ -1,0 +1,6 @@
+import { useMDXComponents as getDocsMDXComponents } from '@theguild/components/server';
+import { docsMDXComponents } from './src/components/docs-mdx-components';
+
+export const useMDXComponents = () => {
+  return getDocsMDXComponents(docsMDXComponents);
+};
