+++ packages/web/docs/tailwind.config.ts
@@ -10,6 +10,7 @@ import baseConfig from '@theguild/tailwind-config';
 
 const config: Config = {
   ...baseConfig,
+  content: [...baseConfig.content, './mdx-components.js'],
   theme: {
     ...baseConfig.theme,
     extend: {
